% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ivLT.R
\name{BrC.r}
\alias{BrC.ds_dphi}
\alias{BrC.r}
\title{Bromwich Contour}
\usage{
BrC.r(phi, gamma = 1)

BrC.ds_dphi(phi, gamma = 1)
}
\arguments{
\item{phi}{\eqn{\phi} value}

\item{gamma}{the Bromwich contour is a straight line and intersects the real axis at \eqn{\gamma}}
}
\description{
The Bromwich contour with polar co-ordinates (\eqn{r} as a function of \eqn{\phi})
}
\section{Functions}{
\itemize{
\item \code{BrC.r}: \eqn{r}

\item \code{BrC.ds_dphi}: \eqn{ds/d\phi}
}}

