% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PolarCart.R
\name{r.xy}
\alias{phi.xy}
\alias{r.xy}
\title{Cartesian to Polar}
\usage{
r.xy(x, y)

phi.xy(x, y)
}
\arguments{
\item{x}{x co-ordinate}

\item{y}{y co-ordinate}
}
\value{
r or phi respectively from x and y
}
\description{
Conversion from cartesian to polar co-ordinates
}
\section{Functions}{
\itemize{
\item \code{r.xy}: Returns polar co-ordinate r from cartesian co-ordinates x and y.

\item \code{phi.xy}: Returns polar co-ordinate phi (anti-clockwise rotation from positive x-axis) from cartesian co-ordinates x and y.
}}

