% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PolarCart.R
\name{x.rphi}
\alias{x.rphi}
\alias{y.rphi}
\title{Polar to Cartesian}
\usage{
x.rphi(r, phi)

y.rphi(r, phi)
}
\arguments{
\item{r}{distance from origin}

\item{phi}{anti-clockwise rotation from positive x-axis}
}
\value{
x or t respectively from r and phi
}
\description{
Conversion from polar to cartesian co-ordinates
}
\section{Functions}{
\itemize{
\item \code{x.rphi}: Returns cartesian co-ordinate x from polar co-ordinates r and phi.

\item \code{y.rphi}: Returns cartesian co-ordinate x from polar co-ordinates r and phi.
}}

